requirements.txt
pytest==7.1.2
pytest-mock==3.7.0
pyyaml==6.0
